refactor: extract chat helpers
